working stream title display, terribly slow

+++ twitchtitle.sh
@@ -3,11 +3,10 @@
 #if [ $# -eq 2 ] ; then
 #output='tmpdl'
 output=$(date +%s.%N)
-#if [ $# -ge 1 ] ; then
+if [ $# -ge 1 ] ; then
 
-    echo 'ok'
     wget $1 -O $output --quiet
     titlestring=$(grep "property='og:description'>" $output)
     echo $titlestring |  sed "s/<meta content='//" | sed "s/' property='og:description'>//"
     rm $output
-#fi
+fi
